# apps/web/app/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LOGOS API",
    "description": "API for classical text analysis, translation quality, and intertextual research",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Local API" },
    { "url": "https://logos-backend-production-0d96.up.railway.app", "description": "Production backend" }
  ],
  "paths": {
...
    "/translations": {
      "get": {
        "summary": "List translations with quality scores",
        "tags": ["Translations"],
        "parameters": [
          { "name": "translator", "in": "query", "schema": { "type": "string" } },
          { "name": "flagged", "in": "query", "schema": { "type": "boolean" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/TranslationQuality" } },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/intertexts": {
      "get": {
        "summary": "Get all intertextual connections for graph visualization",
        "tags": ["Intertexts"],
        "parameters": [
          { "name": "minStrength", "in": "query", "schema": { "type": "number", "default": 0.3 } },
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["quotation", "allusion", "echo", "parallel"] } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IntertextGraph" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TranslationQuality": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "urn": { "type": "string" },
          "translator": { "type": "string" },
          "style": { "type": "string" },
          "sourceText": { "type": "string" },
          "translatedText": { "type": "string" },
          "scores": { "$ref": "#/components/schemas/QualityScores" },
          "flagged": { "type": "boolean" },
          "flagReason": { "type": "string" }
        }
      },
      "QualityScores": {
        "type": "object",
        "properties": {
          "overall": { "type": "number" },
          "semanticFidelity": { "type": "number", "description": "SF - How well meaning is preserved" },
          "registerMatch": { "type": "number", "description": "RM - Tone/formality consistency" },
          "literalness": { "type": "number", "description": "LT - Word-for-word vs interpretive" },
          "readability": { "type": "number", "description": "RD - English fluency score" },
          "styleConsistency": { "type": "number", "description": "SC - Matches translator persona" },
          "translatorBias": { "type": "number", "description": "TB - Genre-conditioned deviation" }
        }
      },
...
      "IntertextGraph": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "label": { "type": "string" },
                "author": { "type": "string" },
                "year": { "type": "integer" },
                "size": { "type": "number" }
              }
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "source": { "type": "string" },
                "target": { "type": "string" },
                "type": { "type": "string" },
                "strength": { "type": "number" }
              }
            }
          }
        }
      },
...
    }
  }
}